requirements.txt
djangorestframework

djangorestframework-simplejwt 